migrate button and withForwardedRef to typescript
