+++ react/components/EXPERIMENTAL_Table/Toolbar/PopoverMenu.tsx
@@ -7,7 +7,7 @@ import React, {
 } from 'react'
 import classnames from 'classnames'
 
-import Button from '../../Button/index.js'
+import Button from '../../Button'
 import { E2ETestable } from '../types'
 
 export default function usePopoverMenu() {
@@ -127,7 +127,7 @@ type ItemProps = {
 
 export type MenuAction = {
   label: string
-  onClick: Function
+  onClick: (event: React.MouseEvent) => void
   toggle?: {
     checked: boolean
     semantic: boolean
