+++ react/modules/withForwardedRef.tsx
@@ -1,10 +1,12 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
+export class Noop {}
+
 // For more info see: https://stackoverflow.com/a/51127130/10725088
 const Element =
   typeof window === 'undefined' || typeof window.Element === 'undefined'
-    ? function() {}
+    ? Noop
     : window.Element
 
 export const refShape = PropTypes.oneOfType([
@@ -14,8 +16,8 @@ export const refShape = PropTypes.oneOfType([
   }),
 ])
 
-export function withForwardedRef(Component) {
-  const ComponentWithRef = React.forwardRef((props, ref) => {
+export function withForwardedRef<T, P>(Component: React.ComponentType<P>) {
+  const ComponentWithRef = React.forwardRef<T, P>((props, ref) => {
     return <Component {...props} forwardedRef={ref} />
   })
